// types/service-type.ts
export enum ServiceType {
  Reconnection = "Re-connection",
  Bowser = "Bowser",
  LeakDetection = "Leak Detection",
  MeterTest = "Meter Test",
  ChangeOfConnection = "Change of Connection",
  SewerUnblocking = "Sewer Unblocking",
  WaterConnection = "Water Connection",
  SewerConnection = "Sewer Connection",
  OnsiteSanitation = "Onsite Sanitation",
  FeacalSludgeMgt = "Feacal Sludge Management",
}